Move AssemblyDependencyUtil to Common.
